refactor: use top-level plugins key and extract .imposter.yaml to own file
Move .imposter.yaml generation logic from configfile.go to a dedicated
dot_imposter.go file. Update the plugins section to use the top-level
`plugins` key instead of the deprecated `default.plugins` nesting.
